# docs/swagger.yaml
basePath: /api
definitions:
  APIError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  APIErrors:
    properties:
      code:
        type: integer
      errors:
        items:
          $ref: '#/definitions/APIError'
        type: array
    type: object
  AddCartItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        minimum: 1
        type: integer
      variant_id:
        type: string
    required:
    - product_id
    - quantity
    - variant_id
    type: object
...
  CartItemResponse:
    properties:
      available:
        type: boolean
      image:
        $ref: '#/definitions/ImageDTO'
      max_quantity:
        type: integer
      name:
        type: string
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
      slug:
        type: string
      variant_id:
        type: string
    type: object
  CartResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/CartItemResponse'
        type: array
      total_price:
        type: number
    type: object
...
  CityResponse:
    properties:
      address:
        type: string
      area:
        type: string
      area_type:
        type: string
      capital_marker:
        type: integer
      city:
        type: string
      city_type:
        type: string
      country:
        type: string
      federal_district:
        type: string
      fias_id:
        type: string
      fias_level:
        type: integer
      foundation_year:
        type: integer
      geo_lat:
        type: number
      geo_lon:
        type: number
      id:
        type: string
      kladr_id:
        type: string
      okato:
        type: string
      oktmo:
        type: string
      population:
        type: integer
      postal_code:
        type: string
      region:
        type: string
      region_type:
        type: string
      settlement:
        type: string
      settlement_type:
        type: string
      tax_office:
        type: string
      timezone:
        type: string
    type: object
...
  ImageDTO:
    properties:
      alt:
        type: string
      height:
        type: integer
      id:
        type: string
      presets:
        additionalProperties:
          additionalProperties:
            type: string
          type: object
        type: object
      width:
        type: integer
    type: object
...
  JSONResponse-CartResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/CartResponse'
      message:
        type: string
    type: object
...
  JSONResponse-CityResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/CityResponse'
      message:
        type: string
    type: object
...
  JSONResponse-any:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  JSONResponse-array_CityResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/CityResponse'
        type: array
      message:
        type: string
    type: object
...
  UpdateCartItemRequest:
    properties:
      quantity:
        type: integer
    type: object
...
info:
  contact:
    email: go-store@stick.sh
    name: Vladislav B
  description: This is an API for go-store
  title: GO-store
  version: "1.0"
paths:
  /api/v1/cart:
    delete:
      description: Removes all items from the cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Clear cart
      tags:
      - Cart
    get:
      description: Returns the current cart for authenticated user or guest session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-CartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Get cart
      tags:
      - Cart
  /api/v1/cart/items:
    post:
      consumes:
      - application/json
      description: Adds a product variant to the cart. If the variant is already present,
        its quantity is increased.
      parameters:
      - description: Item to add
        in: body
        name: Add
        required: true
        schema:
          $ref: '#/definitions/AddCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-CartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Add item to cart
      tags:
      - Cart
  /api/v1/cart/items/{variantId}:
    delete:
      description: Removes the specified variant from the cart
      parameters:
      - description: Variant UUID
        in: path
        name: variantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-CartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Remove item from cart
      tags:
      - Cart
    patch:
      consumes:
      - application/json
      description: Sets a new quantity for the specified variant in the cart
      parameters:
      - description: Variant UUID
        in: path
        name: variantId
        required: true
        type: string
      - description: New quantity
        in: body
        name: create
        required: true
        schema:
          $ref: '#/definitions/UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-CartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Update cart item quantity
      tags:
      - Cart
...
  /v1/geo/city/:
    get:
      consumes:
      - application/json
      description: Get city by IP address
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-CityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Geo city
      tags:
      - Geo
  /v1/geo/city/find:
    get:
      consumes:
      - application/json
      description: Find city by name
      parameters:
      - description: City name
        in: query
        name: city
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-array_CityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Find city
      tags:
      - Geo
  /v1/geo/city/popular:
    get:
      consumes:
      - application/json
      description: Get most popular city
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-CityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Get popular city
      tags:
      - Geo
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"